// src/core/actorTypes.d.ts
export interface BaseMessage {
    type: string;
}